+++ packages/stryker-jasmine-runner/src/JasmineTypes.ts
@@ -1,0 +1,111 @@
+declare namespace jasmine {
+  interface Order {
+    new(options: { random: boolean, seed: string }): any;
+    random: boolean;
+    seed: string;
+    sort<T>(items: T[]): T[];
+  }
+
+  interface Trace {
+    name: string;
+    message: string;
+    stack: any;
+  }
+
+  interface Result {
+    type: string;
+  }
+  interface ExpectationResult extends Result {
+    matcherName: string;
+    passed(): boolean;
+    expected: any;
+    actual: any;
+    message: string;
+    trace: Trace;
+  }
+
+  interface RunDetails {
+    failedExpectations: ExpectationResult[];
+    order: Order;
+  }
+
+
+  interface SuiteInfo {
+    totalSpecsDefined: number;
+  }
+
+  interface CustomReportExpectation {
+    // matcherName: string;
+    message: string;
+    // passed: boolean;
+    // stack: string;
+  }
+
+  interface FailedExpectation extends CustomReportExpectation {
+    // actual: string;
+    // expected: string;
+  }
+
+  interface PassedExpectation extends CustomReportExpectation {
+
+  }
+
+  interface CustomReporterResult {
+    // description: string;
+    failedExpectations?: FailedExpectation[];
+    fullName: string;
+    id: string;
+    passedExpectations?: PassedExpectation[];
+    pendingReason?: string;
+    status?: string;
+  }
+  interface CustomReporter {
+    jasmineStarted?(suiteInfo: SuiteInfo): void;
+    suiteStarted?(result: CustomReporterResult): void;
+    specStarted(/*result: CustomReporterResult // not needed ;) */): void;
+    specDone(result: CustomReporterResult): void;
+    suiteDone?(result: CustomReporterResult): void;
+    jasmineDone(/*runDetails: RunDetails // not needed ;) */): void;
+  }
+
+  interface Env {
+    throwOnExpectationFailure(value: boolean): void;
+  }
+}
+
+declare module 'jasmine' {
+
+  class Jasmine {
+    constructor(options: any);
+    addReporter(reporter: jasmine.CustomReporter): void;
+    addSpecFile(filePath: string): void;
+    addSpecFiles(files: string[]): void;
+    clearReporters(): void;
+    configureDefaultReporter(options: any, ...args: any[]): void;
+    env: jasmine.Env;
+    execute(files?: string[], filterString?: string): any;
+    exit: (code: number) => void;
+    exitCodeCompletion(passed: any): void;
+    loadConfig(config: any): void;
+    loadConfigFile(configFilePath: any): void;
+    loadHelpers(): void;
+    loadSpecs(): void;
+    onComplete(onCompleteCallback: (passed: boolean) => void): void;
+    randomizeTests(value?: any): boolean;
+    seed(value: any): void;
+    showColors(value: any): void;
+    stopSpecOnExpectationFailure(value: boolean): void;
+    stopOnSpecFailure(value: boolean): void;
+    static ConsoleReporter(): any;
+    reportersCount: number;
+    completionReporter: jasmine.CustomReporter;
+    reporter: jasmine.CustomReporter;
+    coreVersion(): string;
+    showingColors: boolean;
+    projectBaseDir: string;
+    printDeprecation(): void;
+    specFiles: string[];
+    helperFiles: string[];
+  }
+  export = Jasmine;
+}