+++ packages/stryker-typescript/testResources/allowJS/useMath.js
@@ -1,0 +1,5 @@
+const math = require('./math');
+
+const a = 42;
+const b = 31;
+console.log(`a + b `, math.add(a, b));