+++ packages/stryker-webpack-transpiler/src/compiler/ConfigLoader.ts
@@ -3,6 +3,7 @@ import * as fs from 'fs';
 import { Configuration } from 'webpack';
 import { StrykerWebpackConfig } from '../WebpackTranspiler';
 import { getLogger, Logger } from 'log4js';
+import { isFunction } from 'lodash';
 
 const PROGRESS_PLUGIN_NAME = 'ProgressPlugin';
 
@@ -15,11 +16,14 @@ export default class ConfigLoader {
     this.log = getLogger(ConfigLoader.name);
   }
 
-  public load(config: StrykerWebpackConfig): Configuration {
+  public async load(config: StrykerWebpackConfig): Promise<Configuration> {
     let webpackConfig: Configuration;
 
     if (config.configFile) {
-      webpackConfig = this.loaderWebpackConfigFromProjectRoot(config.configFile);
+      webpackConfig = await this.loadWebpackConfigFromProjectRoot(config.configFile);
+      if (isFunction(webpackConfig)) {
+        webpackConfig = webpackConfig.apply(null, config.configFileArgs);
+      }
       if (config.silent) {
         this.configureSilent(webpackConfig);
       }
@@ -31,7 +35,7 @@ export default class ConfigLoader {
     return webpackConfig;
   }
 
-private loaderWebpackConfigFromProjectRoot(configFileLocation: string) {
+private loadWebpackConfigFromProjectRoot(configFileLocation: string) {
   const resolvedName = path.resolve(configFileLocation);
 
   if (!fs.existsSync(resolvedName)) {