+++ packages/stryker-webpack-transpiler/stryker.conf.js
@@ -2,7 +2,9 @@ module.exports = function (config) {
   config.set({
     mutate: [
       'src/**/*.ts',
-      '!src/**/*.d.ts'
+      '!src/**/*.d.ts',
+      '!src/enhanced-resolve.ts',
+      '!src/index.ts'
     ],
     testRunner: "mocha",
     testFramework: "mocha",