+++ packages/stryker/src/reporters/ci/CircleProvider.ts
@@ -1,0 +1,21 @@
+import { CIProvider } from './Provider';
+
+import { getEnvironmentVariable } from '../../utils/objectUtils';
+
+class CircleProvider implements CIProvider {
+  isPullRequest = () => getEnvironmentVariable('CIRCLE_PULL_REQUEST') !== undefined;
+
+  determineBranch = () => getEnvironmentVariable('CIRCLE_BRANCH') || '(unknown)';
+
+  determineRepository = () => {
+    const username = getEnvironmentVariable('CIRCLE_PROJECT_USERNAME');
+    const reponame = getEnvironmentVariable('CIRCLE_PROJECT_REPONAME');
+    if (username !== '' && reponame !== '') {
+      return `${username}/${reponame}`;
+    } else {
+      return '(unknown)';
+    }
+  }
+}
+
+export default CircleProvider;