+++ packages/stryker-html-reporter/CHANGELOG.md
@@ -3,6 +3,22 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+<a name="0.14.2"></a>
+## [0.14.2](https://github.com/stryker-mutator/stryker/compare/stryker-html-reporter@0.14.1...stryker-html-reporter@0.14.2) (2018-05-31)
+
+
+
+
+**Note:** Version bump only for package stryker-html-reporter
+
+<a name="0.14.1"></a>
+## [0.14.1](https://github.com/stryker-mutator/stryker/compare/stryker-html-reporter@0.14.0...stryker-html-reporter@0.14.1) (2018-05-21)
+
+
+
+
+**Note:** Version bump only for package stryker-html-reporter
+
 <a name="0.14.0"></a>
 # [0.14.0](https://github.com/stryker-mutator/stryker/compare/stryker-html-reporter@0.13.3...stryker-html-reporter@0.14.0) (2018-04-30)